requirements.txt
alembic==1.13.2
annotated-types==0.7.0
anyio==4.9.0
app==0.0.1
appnope==0.1.3
asttokens==2.4.0
asyncpg==0.30.0
autopep8==2.3.1
backcall==0.2.0
bcrypt==4.2.0
blinker==1.6.2
CacheControl==0.14.0
cachelib==0.13.0
cachetools==5.3.3
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==2.0.12
click==8.1.7
comm==0.1.4
contourpy==1.1.1
coverage==7.3.1
cryptography==42.0.5
css==0.1
customtkinter==5.2.0
cycler==0.12.1
darkdetect==0.8.0
debugpy==1.8.0
decorator==5.1.1
distlib==0.3.7
dnspython==2.6.1
docker==7.1.0
docutils==0.21.2
email-validator==1.1.3
executing==1.2.0
factory_boy==3.0.0
Faker==26.0.0
fastapi==0.115.11
filelock==3.13.1
firebase==4.0.1
firebase-admin==6.5.0
Flask==3.0.3
flask-cors==5.0.1
Flask-Dance==7.1.0
Flask-Login==0.5.0
Flask-Mail==0.10.0
Flask-Migrate==4.0.4
Flask-PyMongo==2.3.0
Flask-Script==2.0.6
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
Flask-Testing==0.8.1
Flask-WTF==0.15.1
fonttools==4.43.1
geographiclib==2.0
geohash2==1.1
geopy==2.4.1
gitdb==4.0.11
GitPython==3.1.43
golang==0.0.3
google-ai-generativelanguage==0.6.6
google-api-core==2.18.0
google-api-python-client==2.125.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.0
google-cloud-speech==2.27.0
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-generativeai==0.7.2
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
googlemaps==4.10.0
grpcio==1.62.1
grpcio-status==1.62.1
h11==0.14.0
httplib2==0.22.0
idna==3.4
iniconfig==2.0.0
ipykernel==6.25.2
ipython==8.15.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
jupyter_client==8.3.1
jupyter_core==5.3.1
kiwisolver==1.4.5
linalg==1.0.4
load-dotenv==0.1.0
lu==0.1.0
lxml==5.3.0
Mako==1.3.5
mangodb==0.9
MarkupSafe==2.1.3
matplotlib==3.4.3
matplotlib-inline==0.1.6
msgpack==1.0.8
msgspec==0.18.6
mysql-connector-python==8.3.0
mysql-connector-python-rf==2.2.2
nest-asyncio==1.5.8
networkx==3.2.1
npm==0.1.1
numpy==1.26.0
oauthlib==3.2.2
ocaml==0.0.11
optional-django==0.1.0
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
pipenv==2023.11.15
platformdirs==3.10.0
pluggy==1.5.0
prompt-toolkit==3.0.39
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.12.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pygame==2.5.2
Pygments==2.16.1
PyJWT==2.8.0
pymongo==4.7.3
pyparsing==3.1.1
pytest==8.3.5
pytest-flask==1.3.0
python-dateutil==2.8.2
python-docx==1.1.2
python-dotenv==0.19.2
pytz==2024.1
pyzmq==25.1.1
redis==5.2.1
requests==2.26.0
requests-oauthlib==2.0.0
rsa==4.9
scikit-learn==1.3.2
scipy==1.11.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
spotipy==2.25.0
SQLAlchemy==2.0.31
stack-data==0.6.2
starlette==0.46.1
threadpoolctl==3.2.0
tornado==6.3.3
tqdm==4.66.4
traitlets==5.10.0
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==1.26.19
URLObject==2.4.3
uvicorn==0.34.0
virtualenv==20.24.7
wcwidth==0.2.6
Werkzeug==3.0.3
WTForms==3.0.1
wurlitzer==3.1.0
